Fix sign problem when specifying a large array index
Requesting an array index > UINT_MAX leads to an unexpected array id:

    $ sbatch --array=4294967297 --wrap=hostname
    Submitted batch job 12853
    $ scontrol show job 12853|grep -i array
    JobId=12853 ArrayJobId=12853 ArrayTaskId=1 JobName=wrap

Since MaxArraySize can't exceed 4000001, this job should be rejected.
This patch fixes that.

Bug 15354

+++ src/slurmctld/job_mgr.c
@@ -7774,7 +7774,7 @@ static int _test_strlen(char *test_str, char *str_name, int max_str_len)
 static bool _parse_array_tok(char *tok, bitstr_t *array_bitmap, uint32_t max)
 {
 	char *end_ptr = NULL;
-	int i, first, last, step = 1;
+	long int i, first, last, step = 1;
 
 	if (tok[0] == '[')	/* Strip leading "[" */
 		tok++;
